+++ Altinn/AT.Common.Altinn.Publish/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security <!-- in case of vulnerabilities. -->
 
+## 3.1.1
+
+### Fixed
+
+- fix(altinn): added necessary null checks when converting an altinn instance to metadata
+
 ## 3.1.0
 
 ### Added